[MOVE/#78] 추천친구뷰 패키지 세분화
