+++ app/src/main/java/com/el/yello/presentation/main/recommend/RecommendViewPagerAdapter.kt
@@ -2,6 +2,8 @@ package com.el.yello.presentation.main.recommend
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.el.yello.presentation.main.recommend.kakao.RecommendKakaoFragment
+import com.el.yello.presentation.main.recommend.school.RecommendSchoolFragment
 
 class RecommendViewPagerAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
